+++ src/DemonsInBetween.cpp
@@ -8,41 +8,48 @@
 
 using namespace geode::prelude;
 
-LadderDemon& DemonsInBetween::demonForLevel(int levelID, bool main) {
+LadderDemon& DemonsInBetween::demonForLevel(int levelID) {
     static LadderDemon empty = { 0, 0.0, 0.0, 0 };
-    auto& gddl = main ? GDDL_MAIN : GDDL;
-    auto demon = ranges::indexOf(gddl, [levelID, main](const LadderDemon& d) {
-        return d.id == levelID;
-    });
-    return !demon.has_value() ? empty : gddl[demon.value()];
+    auto demon = ranges::indexOf(GDDL, [levelID](const LadderDemon& d) { return d.id == levelID; });
+    return !demon.has_value() ? empty : GDDL[demon.value()];
 }
 
-CCSpriteFrame* DemonsInBetween::spriteFrameForDifficulty(int difficulty, GJDifficultyName name, GJFeatureState state) {
-    return CCSpriteFrameCache::get()->spriteFrameByName(fmt::format(
+CCPoint DemonsInBetween::offsetForDifficulty(int difficulty, GJDifficultyName name) {
+    switch (difficulty) {
+        case 1: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -5.0f } : CCPoint { -0.125f, -0.25f };
+        case 2: return name == GJDifficultyName::Long ? CCPoint { 0.125f, -5.0f } : CCPoint { -0.125f, -0.25f };
+        case 3: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -5.0f } : CCPoint { -0.125f, -0.25f };
+        case 4: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -5.125f } : CCPoint { -0.125f, -0.375f };
+        case 5: return name == GJDifficultyName::Long ? CCPoint { 0.25f, -5.0f } : CCPoint { -0.125f, -0.25f };
+        case 6: return name == GJDifficultyName::Long ? CCPoint { 0.125f, -4.75f } : CCPoint { -0.125f, -0.25f };
+        case 7: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -5.0f } : CCPoint { -0.125f, -0.375f };
+        case 8: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -4.125f } : CCPoint { -0.125f, 0.5f };
+        case 9: return name == GJDifficultyName::Long ? CCPoint { -0.125f, -4.125f } : CCPoint { -0.125f, 0.5f };
+        case 10: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -4.0f } : CCPoint { -0.125f, 0.25f };
+        case 11: return name == GJDifficultyName::Long ? CCPoint { -0.125f, -4.125f } : CCPoint { -0.125f, 0.5f };
+        case 12: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -4.125f } : CCPoint { 0.125f, 0.5f };
+        case 13: return name == GJDifficultyName::Long ? CCPoint { 0.125f, -4.125f } : CCPoint { 0.125f, 0.5f };
+        case 14: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -4.125f } : CCPoint { 0.125f, 0.5f };
+        case 15: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -4.125f } : CCPoint { 0.0f, 0.5f };
+        case 16: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -3.625f } : CCPoint { 0.0f, 1.25f };
+        case 17: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -3.625f } : CCPoint { 0.0f, 1.25f };
+        case 18: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -3.5f } : CCPoint { 0.0f, 1.125f };
+        case 19: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -3.5f } : CCPoint { 0.0f, 1.125f };
+        case 20: return name == GJDifficultyName::Long ? CCPoint { 0.0f, -3.5f } : CCPoint { 0.0f, 1.125f };
+        default: return CCPoint { 0.0f, 0.0f };
+    }
+}
+
+CCSprite* DemonsInBetween::spriteForDifficulty(GJDifficultySprite* difficultySprite, int difficulty, GJDifficultyName name, GJFeatureState state) {
+    auto sprite = CCSprite::createWithSpriteFrameName(fmt::format(
         "DIB_{:02d}{}_btn{}_001.png"_spr,
         difficulty,
         state == GJFeatureState::Legendary ? "_4" : state == GJFeatureState::Mythic ? "_5" : "",
         name == GJDifficultyName::Long ? "2" : ""
     ).c_str());
-}
-
-CCSprite* DemonsInBetween::spriteForDifficulty(GJDifficultySprite* difficultySprite, int difficulty, GJDifficultyName name, GJFeatureState state) {
-    auto sprite = CCSprite::createWithSpriteFrame(spriteFrameForDifficulty(difficulty, name, state));
-    sprite->setPosition(difficultySprite->getPosition() + (name == GJDifficultyName::Long ?
-        LONG_OFFSETS[(size_t)difficulty - 1] : SHORT_OFFSETS[(size_t)difficulty - 1]));
+    sprite->setPosition(difficultySprite->getPosition() + offsetForDifficulty(difficulty, name));
     sprite->setID("between-difficulty-sprite"_spr);
     return sprite;
-}
-
-int DemonsInBetween::difficultyForDemonDifficulty(int demonDifficulty) {
-    switch (demonDifficulty) {
-        case 0: return 11;
-        case 3: return 4;
-        case 4: return 7;
-        case 5: return 15;
-        case 6: return 20;
-        default: return 0;
-    }
 }
 
 GJFeatureState DemonsInBetween::stateForLevel(GJGameLevel* level) {
@@ -52,77 +59,10 @@ GJFeatureState DemonsInBetween::stateForLevel(GJGameLevel* level) {
     return state;
 }
 
-void DemonsInBetween::loadDemonForLevel(
-    EventListener<web::WebTask>&& listenerRef, int levelID, bool main,
-    const std::function<void(LadderDemon&)>& callback
-) {
-    if (LEVELS_LOADED.contains(levelID)) return;
-
-    auto&& listener = std::move(listenerRef);
-    listener.bind([callback, levelID, main](web::WebTask::Event* e) {
-        if (auto res = e->getValue()) {
-            LEVELS_LOADED.insert(levelID);
-
-            if (!res->ok()) return;
-
-            auto json = res->json().unwrapOr(matjson::Value());
-            if (!json.contains("Rating") || json["Rating"].isNull()) return;
-
-            auto rating = round(json["Rating"].asDouble().unwrapOr(0.0) * 100) / 100;
-            auto difficulty = DemonsInBetween::DIFFICULTY_INDICES[(int)round(rating)];
-            auto enjoyment = !json.contains("Enjoyment") || json["Enjoyment"].isNull() ? -999.0 : round(json["Enjoyment"].asDouble().unwrapOr(-999.0) * 100) / 100;
-
-            auto& gddl = main ? GDDL_MAIN : GDDL;
-            gddl.push_back({ levelID, rating, enjoyment, difficulty });
-            callback(demonForLevel(levelID, main));
-        }
-    });
-
-    listener.setFilter(web::WebRequest().get(fmt::format("https://gdladder.com/api/level/{}", levelID)));
-}
-
-void DemonsInBetween::searchObjectForPage(
-    EventListener<web::WebTask>&& listenerRef, int page, bool refresh,
-    const std::function<void(GJSearchObject*)>& callback
-) {
-    auto glm = GameLevelManager::get();
-    auto searchCache = static_cast<CCDictionary*>(glm->getUserObject("search-cache"_spr));
-    if (!searchCache) {
-        searchCache = CCDictionary::create();
-        glm->setUserObject("search-cache"_spr, searchCache);
-    }
-
-    auto searchKey = fmt::format("{}_{}", DIFFICULTY, page);
-    if (!refresh && searchCache->objectForKey(searchKey)) {
-        SEARCH_SIZE = SEARCH_SIZES[DIFFICULTY];
-        MAX_PAGE = (SEARCH_SIZE - 1) / 10;
-        return callback(static_cast<GJSearchObject*>(searchCache->objectForKey(searchKey)));
-    }
-
-    auto tierBound = TIER_BOUNDS[DIFFICULTY];
-
-    auto&& listener = std::move(listenerRef);
-    listener.bind([callback, page](web::WebTask::Event* e) {
-        if (auto res = e->getValue()) {
-            if (!res->ok()) return;
-
-            auto json = res->json().unwrapOr(matjson::Value());
-            SEARCH_SIZE = json["total"].asInt().unwrapOr(0);
-            SEARCH_SIZES[DIFFICULTY] = SEARCH_SIZE;
-            MAX_PAGE = (SEARCH_SIZE - 1) / 10;
-
-            auto searchObject = GJSearchObject::create(SearchType::MapPackOnClick, json.contains("levels") && json["levels"].isArray() ? string::join(
-                ranges::map<std::vector<std::string>>(ranges::filter(json["levels"].asArray().unwrap(), [](const matjson::Value& level) {
-                    return level.contains("ID") && level["ID"].isNumber();
-                }), [](const matjson::Value& level) {
-                    return std::to_string(level["ID"].asInt().unwrap());
-                }), ",") : "");
-            if (auto searchCache = static_cast<CCDictionary*>(GameLevelManager::get()->getUserObject("search-cache"_spr)))
-                searchCache->setObject(searchObject, fmt::format("{}_{}", DIFFICULTY, page));
-            callback(searchObject);
-        }
-    });
-
-    listener.setFilter(web::WebRequest().get(fmt::format("https://gdladder.com/api/level/search/?lowTier={}&highTier={}&chunk=10&page={}",
-        tierBound.first, tierBound.second, page + 1)));
+GJSearchObject* DemonsInBetween::searchObjectForPage(int page) {
+    auto& levels = DemonsInBetween::GDDL_DIFFICULTIES[DemonsInBetween::DIFFICULTY];
+    SEARCH_SIZE = levels.size();
+    MAX_PAGE = (SEARCH_SIZE - 1) / 10;
+    return GJSearchObject::create(SearchType::MapPackOnClick, string::join(
+        std::vector<std::string>(levels.begin() + page * 10, levels.begin() + std::min(SEARCH_SIZE, (page + 1) * 10)), ","));
 }