+++ src/classes/DIBSearchPopup.cpp
@@ -17,7 +17,13 @@ DIBSearchPopup* DIBSearchPopup::create() {
 }
 
 bool DIBSearchPopup::setup() {
+    setID("DIBSearchPopup");
     setTitle("Quick Search");
+    m_title->setID("quick-search-title");
+    m_mainLayer->setID("main-layer");
+    m_buttonMenu->setID("button-menu");
+    m_bgSprite->setID("background");
+    m_closeBtn->setID("close-button");
     m_noElasticity = true;
 
     auto table = TableNode::create(5, 4);
@@ -25,20 +31,19 @@ bool DIBSearchPopup::setup() {
     table->setColumnLayout(ColumnLayout::create()->setAxisReverse(true));
     table->setRowLayout(RowLayout::create()->setAxisAlignment(AxisAlignment::Even));
     table->setRowHeight(60.0f);
+    table->setRowPrefix("search-button-row");
     table->setPosition({ 175.0f, 130.0f });
+    table->setID("search-buttons");
     m_mainLayer->addChild(table);
 
     for (int i = 1; i < 21; i++) {
-        table->addButton(CCMenuItemExt::createSpriteExtraWithFrameName(fmt::format("DIB_{:02d}_btn2_001.png"_spr, i).c_str(), 1.0f, [this, i](auto) {
-            if (m_isBusy) return;
-            m_isBusy = true;
+        auto button = CCMenuItemExt::createSpriteExtraWithFrameName(fmt::format("DIB_{:02d}_btn2_001.png"_spr, i).c_str(), 1.0f, [this, i](auto) {;
             DemonsInBetween::DIFFICULTY = i;
             DemonsInBetween::SEARCHING = true;
-            DemonsInBetween::searchObjectForPage(std::move(m_listener), 0, false, [this](GJSearchObject* obj) {
-                CCDirector::get()->pushScene(CCTransitionFade::create(0.5f, LevelBrowserLayer::scene(obj)));
-                m_isBusy = false;
-            });
-        }));
+            CCDirector::get()->pushScene(CCTransitionFade::create(0.5f, LevelBrowserLayer::scene(DemonsInBetween::searchObjectForPage(0))));
+        });
+        button->setID(fmt::format("search-button-{}", i).c_str());
+        table->addButton(button);
     }
 
     table->updateAllLayouts();