+++ src/classes/TableNode.cpp
@@ -44,6 +44,14 @@ void TableNode::setRowHeight(float rowHeight) {
     }
 }
 
+void TableNode::setRowPrefix(const std::string& rowPrefix) {
+    m_rowPrefix = rowPrefix;
+    for (int i = 0; i < m_menus->count(); i++) {
+        auto menu = static_cast<CCMenu*>(m_menus->objectAtIndex(i));
+        menu->setID(fmt::format("{}-{}", rowPrefix, i + 1));
+    }
+}
+
 void TableNode::updateAllLayouts() {
     for (auto menu : CCArrayExt<CCMenu*>(m_menus)) {
         menu->updateLayout();
@@ -57,6 +65,7 @@ void TableNode::addButton(CCMenuItem* button) {
         menu = CCMenu::create();
         menu->setContentSize({ m_obContentSize.width, m_rowHeight });
         menu->setLayout(m_rowLayout);
+        menu->setID(fmt::format("{}-{}", m_rowPrefix, m_menus->count() + 1));
         addChild(menu);
         m_menus->addObject(menu);
     } else menu = static_cast<CCMenu*>(m_menus->objectAtIndex(m_menus->count() - 1));