+++ src/classes/TableNode.hpp
@@ -8,6 +8,7 @@ protected:
     int m_columns;
     int m_rows;
     float m_rowHeight;
+    std::string m_rowPrefix;
 
     bool init(int columns, int rows);
 public:
@@ -16,6 +17,7 @@ public:
     void setColumnLayout(geode::AxisLayout*);
     void setRowLayout(geode::AxisLayout*);
     void setRowHeight(float);
+    void setRowPrefix(const std::string&);
     void updateAllLayouts();
     void addButton(cocos2d::CCMenuItem*);