+++ src/hooks/LevelPage.cpp
@@ -21,14 +21,14 @@ class $modify(DIBLevelPage, LevelPage) {
         auto levelID = level->m_levelID.value();
         if (levelID < 1 || GameStatsManager::get()->getStat("8") < m_level->m_requiredCoins) return;
 
-        auto demon = DemonsInBetween::demonForLevel(levelID, true);
+        auto& demon = DemonsInBetween::demonForLevel(levelID);
         if (demon.id == 0 || demon.difficulty == 0) return;
 
         auto overcharged = Loader::get()->getLoadedMod("firee.overchargedlevels");
         auto overchargedEnabled = overcharged && overcharged->getSettingValue<bool>("enabled");
         auto demonSprite = CCSprite::createWithSpriteFrameName(fmt::format("DIB_{:02d}{}_001.png"_spr, demon.difficulty, overchargedEnabled ? "_btn" : "").c_str());
         demonSprite->setPosition(m_difficultySprite->getPosition() +
-            (overchargedEnabled ? DemonsInBetween::SHORT_OFFSETS[(size_t)demon.difficulty - 1] * 0.9f : CCPoint { 0.0f, 0.0f }));
+            (overchargedEnabled ? DemonsInBetween::offsetForDifficulty(demon.difficulty, GJDifficultyName::Short) * 0.9f : CCPoint { 0.0f, 0.0f }));
         demonSprite->setScale(overchargedEnabled ? 0.9f : 1.1f);
         demonSprite->setID("between-difficulty-sprite"_spr);
         m_levelDisplay->addChild(demonSprite);